server/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
daphne
gunicorn
whitenoise
Pillow
python-dotenv
google-auth
requests
boto3
django-storages
stripe

dj-database-url
psycopg2-binary
uvicorn[standard]

# Development / Seeding only
Faker
